v1.3.0 (Fix): Updated README, added `hash_current_passwords()` back to `__init__.py`
